Prepare crate for preview publishing.
Specify License as MIT OR Apache-2.0.
Add license files.

Rename the crate to spmc_queue.

Use push()/pop() instead of enqueue()/dequeue().

Implement Debug, Display and Default.

Implement From to convert a Queue to a Consumer.

Implement new() method, which returns a queue with capacity of 1024.

Add tests for the new functionality.